+++ src/prefect_server/api/deprecated/__init__.py
@@ -1,1 +1,0 @@
-import prefect_server.api.deprecated.flows